# System.Text.Json: .NET 5.0 perf regression serializing Nullable<T>