--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterOfT.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonConverterOfT.cs
@@ -294,6 +294,8 @@ namespace System.Text.Json.Serialization
             return success;
         }
 
+        internal virtual bool IsNull(T value) => value == null;
+
         internal bool TryWrite(Utf8JsonWriter writer, in T value, JsonSerializerOptions options, ref WriteStack state)
         {
             if (writer.CurrentDepth >= options.EffectiveMaxDepth)
@@ -342,7 +344,7 @@ namespace System.Text.Json.Serialization
                     }
                 }
             }
-            else if (value == null && !HandleNullOnWrite)
+            else if (IsNull(value) && !HandleNullOnWrite)
             {
                 // We do not pass null values to converters unless HandleNullOnWrite is true. Null values for properties were
                 // already handled in GetMemberAndWriteJson() so we don't need to check for IgnoreNullValues here.