@steveharter 2021-01-14
It appears the culprit is the `in` modifier + `Nullable<T>` + `==` operator.

When using `==` boxing occurs and when not using `in` a copy is made via `ldobj` probably because `Nullable<T>` is not adorned with the `readonly` modifier. I'll look at creating an issue for `Nullable` once I investigate a bit more.

The fastest approach for `Nullable<T>` is using the `in` + the `Equals()` method. However, since we don't want to call `Equals()` for reference types (since they could overload perhaps and have do something weird that makes semantics inconsistent) keeping your `IsNull()` approach but adding the `in` modifier seems like the best route.

However it would be good to check if a custom value type (without readonly modifier) would be slow here; perhaps a call to `CanBeNull()` would be better before calling `IsNull()` to prevent the method call?

<details>
<summary>Simple repro (click to expand)</summary>

```cs
    class Program
    {
        const long Iterations = 100_000_000;

        public static void Main()
        {
            var converter = new NullableConverter<int>();

            // warm-up (ignore results here)
            EqualsOperator(converter);

            EqualsOperator(converter);
            EqualsOperator_In(converter);
            IsOperator(converter);
            IsOperator_In(converter);
            EqualsMethod(converter);
            EqualsMethod_In(converter);
            VirtualMethod(converter);
            VirtualMethod_In(converter);
        }

        public static void EqualsOperator(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.EqualsOperator(value);
            }
            sw.Stop();

            Console.WriteLine($"EqualsOperator {sw.ElapsedMilliseconds}");
        }

        public static void EqualsOperator_In(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.EqualsOperator_In(value);
            }
            sw.Stop();

            Console.WriteLine($"EqualsOperator_In {sw.ElapsedMilliseconds}");
        }

        public static void IsOperator(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.IsOperator(value);
            }
            sw.Stop();

            Console.WriteLine($"IsOperator {sw.ElapsedMilliseconds}");
        }

        public static void IsOperator_In(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.IsOperator_In(value);
            }
            sw.Stop();

            Console.WriteLine($"IsOperator_In {sw.ElapsedMilliseconds}");
        }

        public static void EqualsMethod(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.EqualsMethod(value);
            }
            sw.Stop();

            Console.WriteLine($"EqualsMethod {sw.ElapsedMilliseconds}");
        }

        public static void EqualsMethod_In(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.EqualsMethod_In(value);
            }
            sw.Stop();

            Console.WriteLine($"EqualsMethod_In {sw.ElapsedMilliseconds}");
        }

        public static void VirtualMethod(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.VirtualMethod(value);
            }
            sw.Stop();

            Console.WriteLine($"VirtualMethod {sw.ElapsedMilliseconds}");
        }

        public static void VirtualMethod_In(JsonConverter<Nullable<int>> converter)
        {
            int? value = 1;

            var sw = new Stopwatch();
            sw.Start();

            for (long i = 0; i < Iterations; i++)
            {
                converter.VirtualMethod_In(value);
            }
            sw.Stop();

            Console.WriteLine($"VirtualMethod_In {sw.ElapsedMilliseconds}");
        }

        public abstract class JsonConverter<T>
        {
            public void EqualsOperator(T value)
            {
                if (value == null)
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void EqualsOperator_In(in T value)
            {
                if (value == null)
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void IsOperator(T value)
            {
                if (value is null)
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void IsOperator_In(in T value)
            {
                if (value is null)
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void EqualsMethod(T value)
            {
                if (value.Equals(null))
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void EqualsMethod_In(in T value)
            {
                if (value.Equals(null))
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void VirtualMethod(T value)
            {
                if (IsNull(value))
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public void VirtualMethod_In(in T value)
            {
                if (IsNull(value))
                {
                    throw new Exception("shouldn't get here");
                }
            }

            public virtual bool IsNull(T value) => throw new Exception("shouldn't be called");
        }

        public sealed class NullableConverter<T> : JsonConverter<T?> where T : struct
        {
            public override bool IsNull(T? value) => !value.HasValue;
        }
    }
 ```

</details>

with results:
```
EqualsOperator 528
EqualsOperator_In 4015
IsOperator 720
IsOperator_In 4008
EqualsMethod 554
EqualsMethod_In 210
VirtualMethod 652
VirtualMethod_In 552
```